Precompute static shader math on the CPU for EffectStack.
